Refactor project-project_type association
The relation between project and project_type is now directly modelled
with project belongs_to a project_type. This means, there is a
foreign_key in the projects table. The join table of the previous
association is deleted.

Furthermore, the manuell syncronisation between mutual data is no longer
necessary since the data query is delegated to project_types where
appropriate.