Change ProjectCustomField
ProjectCustomField class is now dependent on ProjectType class. That is,
when an administrator assigns the ProjectCustomField to one or many
ProjectType the custom field will be renderned in a Project only if
it has the respective project_type_id.

For any unknown reason, the tests failed due to some errors. This commit
should help to find the reason with the diff tool.