Added test patches, update docs
Test patches for ProjectCustomField class consider that some view
snippets are removed or replaced with the deface gem. Redmine core
tests, however, expect the snippets. Thus core tests fails.

Readme, Changelog, and init files are updated to reflect the latest
version 3.0.0.

It is a major update since ProjectCustomField will only
be displayed in a project if the user has assigned the projects
ProjectType to the ProjectCustomField.