Fix load order for project custom fields
Project custom fields should be filtered by its project_types_ids. The
corresponding patch was loaded after Project class. Therefore, the
filter was not active.

This is changed now. The respective test is adapted to reflect the case
where several ProjectCustomField exist but are not all assigned to the
project under review.