Add project custom field validation
Project custom fields will be validated at three places:

i)   Project Types
ii)  Project Custom Fields
iii) Projects

The validation for i), ii) checks whether the underlying project type
has associated projects. If so, the action is prohibited.

The validation for iii) is triggered when the user tries to change
the project type. If the project has no hosts and guests the content
of project custom fields will be deleted!

Alterntively, there is a method implemented but not used, which blocks
the action until the user deletes the content by himself.