Save custom field projects by hidden field
When a project without project type is configured with certain custom
fields enabled then it will lose this configuration as soon as the custom
field itself is configured and saved.

This should be avoided by adding a hidden field for all projects having
no project type but have had the respective custom field enabled.

Implements #381