Save trackers of projects without project type
When a project without project type is configured with certain trackers
enabled then it will lose this configuration as soon as the tracker
itself is configured and saved.

This should be avoided by adding a hidden field for all projects having
no project type but have had the respective tracker enabled.

Implements #382