Add blank line to project type format
Using the custom field of type project type requires to either select one or
several project types or do nothing at all.

Once the user decided to use the field, she has to select always at least one
project type. It is not possible to deselect everything.

This behaviour is changed by adding a blank line on top of the list of project
type masters.

Implements #511